% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeConsensusClusters.R
\name{mergeConsensusClusters}
\alias{mergeConsensusClusters}
\title{mergeConsensusClusters}
\usage{
mergeConsensusClusters(object, Zthr = 2)
}
\arguments{
\item{object}{A object of "getConsensusCluster" function result}

\item{Zthr}{below this z-value two clusters will be merged}
}
\value{
A list of samples' labels with same connectivity. 
Additional attributes include:
\describe{
  \item{Mean_in}{The average Co-association in the cluster}
  \item{Mean_out}{The average co-association outside the cluster}
  \item{SD_in}{The standard deviation of the  Co-association in the cluster}
  \item{SD_out}{The standard deviation of the  Co-association outside the cluster}
  \item{Quality}{The quality of each cluster}
}
}
\description{
This function merge clusters that share consensus.
}
\examples{
\donttest{
library("mlbench")
data(Sonar)

Sonar$Class <- as.numeric(Sonar$Class)
Sonar$Class[Sonar$Class == 1] <- 0 
Sonar$Class[Sonar$Class == 2] <- 1

ClustStab <- clusterStability(data=Sonar, clustermethod=kmeansCluster, dimenreducmethod="UMAP",
                              n_components = 3,featureselection="yes", outcome="Class",
                              fs.pvalue = 0.05,randomTests = 100,trainFraction = 0.7,center=3)

clusterLabels <- getConsensusCluster(ClustStab,who="training",thr=seq(0.80,0.0,-0.1))
NewclusterLabels <- getConsensusCluster(clusterLabels)
}
}
