% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsensusCluster.R
\name{NC_UMAP_Reclassifier}
\alias{NC_UMAP_Reclassifier}
\title{NC_UMAP_Reclassifier reclassification function}
\usage{
NC_UMAP_Reclassifier(classID, data, n_components = 3)
}
\arguments{
\item{classID}{A returned object of ConsesusLables function}

\item{data}{The data used to extract the clusters}

\item{n_components}{the number of components of the UMAP transform}
}
\value{
the label of each sample in the data as ConsensusLabel Object
}
\description{
This function will reclassify the labels based on Nearest Centroid(NC) and UMAP Dimension reduction method
}
