% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStability.r
\name{plot.ClusterStability}
\alias{plot.ClusterStability}
\title{ClusterStability plot function}
\usage{
\method{plot}{ClusterStability}(clusResult, ...)
}
\arguments{
\item{\dots}{Additional arguments passed to the boxplot function.}

\item{object}{A returned object of ClusterStability function}
}
\value{
the box plots data
}
\description{
This function box plots the indexes of the cluster evaluation.
}
