% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsensusCluster.R
\name{plot.ConsesusLables}
\alias{plot.ConsesusLables}
\title{ConsesusLables plot function}
\usage{
\method{plot}{ConsesusLables}(LablesResult, ...)
}
\arguments{
\item{\dots}{Additional arguments passed to the heatmap plot function.}

\item{object}{A returned object of ConsesusLables function}
}
\value{
the heatmap and barplot
}
\description{
This function  plots the heat map of the consensus matrix.
}
