% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyWrapper.R
\name{predict.FuzzyCluster}
\alias{predict.FuzzyCluster}
\title{FuzzyCluster prediction function}
\usage{
\method{predict}{FuzzyCluster}(object, ...)
}
\arguments{
\item{cls}{A returned object of FuzzyCluster function}

\item{testData}{A data set of new samples}
}
\value{
A list of cluster labels
}
\description{
This function predicts the labels of the cluster for new data based on
cluster labels of the training set.
}
\examples{
pre <- predict(cls,testData[,1:4])

}
