% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanShiftWrapper.R
\name{predict.MeanShiftCluster}
\alias{predict.MeanShiftCluster}
\title{MeanShiftCluster prediction function}
\usage{
\method{predict}{MeanShiftCluster}(clusResult, ...)
}
\arguments{
\item{\dots}{New samples set}

\item{object}{A returned object of MeanShiftCluster function}
}
\value{
A list of cluster labels
}
\description{
This function predicts the labels of the cluster for new data based on
cluster labels of the training set.
}
