% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNEwrapper.R
\name{predict.tsneReductor}
\alias{predict.tsneReductor}
\title{tsneReductor prediction function}
\usage{
\method{predict}{tsneReductor}(object, k = NULL, ...)
}
\arguments{
\item{object}{A returned object of tsneReductor function}

\item{k}{The number is used for computing the means of #neighbors with min distance
(#Neighbor=sqrt(#Samples/k).}

\item{\dots}{New samples set}
}
\value{
tsneY:An embedding of new data
}
\description{
This function performs an embedding of new data using an existing embedding.
}
\examples{
\dontrun{
# Dataset
# https://archive.ics.uci.edu/ml/machine-learning-databases/00537/

Data <- read.csv("~/sobar-72.csv")

rndSamples <- sample(nrow(Data),50)
trainData <- Data[rndSamples,]
testData <- Data[-rndSamples,]

tsne_trainData <- tsneReductor(trainData[,1:19],dim = 3,perplexity = 5,max_iter = 1000)

tsne_testData <- predict(tsne_trainData,k=3,testData[,1:19])
}
}
