% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStability.r
\name{summary.ClusterStability}
\alias{summary.ClusterStability}
\title{ClusterStability summary function}
\usage{
\method{summary}{ClusterStability}(clusResult)
}
\arguments{
\item{object}{A returned object of ClusterStability function}
}
\value{
the table summarizing the mean and standard deviations of the clustering
}
\description{
This function prints the main features of the cluster evaluation.
}
