% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsensusCluster.R
\name{summary.ConsesusLables}
\alias{summary.ConsesusLables}
\title{ConsesusLables summary function}
\usage{
\method{summary}{ConsesusLables}(object)
}
\arguments{
\item{object}{A returned object of ConsesusLables function}
}
\value{
the table summarizing the mean and standard deviations of the clustering
}
\description{
This function prints the main statistics of the cluster consensus.
}
