% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStability.r
\name{clusterStability}
\alias{clusterStability}
\title{clustering stability function}
\usage{
clusterStability(
  data = NULL,
  clustermethod = NULL,
  dimenreducmethod = NULL,
  n_components = 3,
  perplexity = 25,
  max_iter = 1000,
  k_neighbor = 3,
  featureselection = NULL,
  outcome = NULL,
  fs.pvalue = 0.05,
  randomTests = 20,
  trainFraction = 0.5,
  pac.thr = 0.1,
  ...
)
}
\arguments{
\item{data}{A Data set}

\item{k}{The number of Clusters}
}
\value{
A list of cluster labels and a R object of class "fcm {ppclust}"
}
\description{
This function computes the clustering stability that helps to select the best number of clusters.
 Feature selection and dimensionality reduction methods can be used before clustering 
 the data.
}
\examples{
library(datasets)
data(iris)

rndSamples <- sample(nrow(iris),100)
trainData <- iris[rndSamples,]
testData <- iris[-rndSamples,]

cls <- FuzzyCluster(trainData[,1:4],3)
}
