% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsensusCluster.R
\name{getConsensusCluster}
\alias{getConsensusCluster}
\title{Consensus Clustering Results}
\usage{
getConsensusCluster(object, who = "training", thr = seq(0.8, 0.3, -0.1))
}
\arguments{
\item{object}{A object of "clusterStability" function result}

\item{who}{This value shows the consensus clustering result of training and testing sets. If who="training" for training set, otherwise other sets.}
}
\value{
A list of samples' labels with same connectivity.
}
\description{
This function gets the labels of the subjects that share the same connectivity.
}
\examples{
Labscluster <- getConsensusCluster(ClustStab,who="training")

}
