% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNEwrapper.R
\name{predict.tsneReductor}
\alias{predict.tsneReductor}
\title{tsneReductor prediction function}
\usage{
\method{predict}{tsneReductor}(object, k = 3, ...)
}
\arguments{
\item{object}{A returned object of tsneReductor function}

\item{k}{The number is used for computing the means of #neighbors with min distance
(#Neighbor=sqrt(#Samples/k).}

\item{testData}{A data set of new samples}
}
\value{
tsneY:An embedding of new data
}
\description{
This function performs an embedding of new data using an existing embedding.
}
\examples{
library(datasets)
data(iris)

rndSamples <- sample(nrow(iris),100)
trainData <- iris[rndSamples,]
testData <- iris[-rndSamples,]

tsne_trainData <- tsneReductor(trainData,dim = 2,perplexity = 25,max_iter = 1000)
tsne_testData <- predict(tsne_trainData,k=3,testData[,1:4])

}
